pyproject.toml
[project]
name = "multi-modal-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-cloud-aiplatform>=1.138.0",
    "google-genai>=1.64.0",
    "markitdown[all]>=0.1.5",
    "numpy>=2.4.2",
    "python-dotenv>=1.2.1",
]

[project.scripts]
multi-modal-rag = "multi_modal_rag:main"

[tool.uv]
package = true